+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Upgrade stripes-components dependency to v2.0.3. Fixes UIU-423.
 * Modal users-in-users app can now search again, thanks to the STCOM-226 fix. Fixes UIU-426.
 * Prevent renewal if new calculated due date is less than or equal to current due date. Fixes UIU-429.
+* Handle renew from current due date. Fixes UIU-428.
 
 ## [2.12.0](https://github.com/folio-org/ui-users/tree/v2.12.0) (2017-11-28)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.11.0...v2.12.0)